bocco-emo/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
emo-platform-api-sdk==1.0.3
fire==0.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.34.0
pydantic==1.10.18
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
